pre-refactor: A few type and organization tweaks (#325)
* Update openai types


* Drop duplicate EvidenceObj and Evidence types from detector.text_contents


* Move UNSUITABLE_INPUT_MESSAGE and UNSUITABLE_OUTPUT_MESSAGE


* Add common module, move apply_masks, get_chunker_ids, slice_codepoints to common.utils


---------
