+++ src/clients/openai.rs
@@ -34,6 +34,7 @@ use crate::{
     config::ServiceConfig,
     health::HealthCheckResult,
     models::{DetectionWarningReason, DetectorParams},
+    orchestrator,
 };
 
 const DEFAULT_PORT: u16 = 8080;
@@ -76,9 +77,8 @@ impl OpenAiClient {
         headers: HeaderMap,
     ) -> Result<ChatCompletionsResponse, Error> {
         let url = self.inner().endpoint(CHAT_COMPLETIONS_ENDPOINT);
-        let stream = request.stream.unwrap_or_default();
         info!("sending Open AI chat completion request to {}", url);
-        if stream {
+        if request.stream {
             let (tx, rx) = mpsc::channel(32);
             let mut event_stream = self
                 .inner()
@@ -106,7 +106,7 @@ impl OpenAiClient {
                                     code: StatusCode::INTERNAL_SERVER_ERROR,
                                     message: format!("deserialization error: {e}"),
                                 };
-                                let _ = tx.send(Err(error)).await;
+                                let _ = tx.send(Err(error.into())).await;
                             }
                         },
                         Err(error) => {
@@ -115,7 +115,7 @@ impl OpenAiClient {
                                 code: StatusCode::INTERNAL_SERVER_ERROR,
                                 message: error.to_string(),
                             };
-                            let _ = tx.send(Err(error)).await;
+                            let _ = tx.send(Err(error.into())).await;
                         }
                     }
                 }
@@ -165,7 +165,7 @@ impl HttpClientExt for OpenAiClient {
 #[derive(Debug)]
 pub enum ChatCompletionsResponse {
     Unary(Box<ChatCompletion>),
-    Streaming(mpsc::Receiver<Result<Option<ChatCompletionChunk>, Error>>),
+    Streaming(mpsc::Receiver<Result<Option<ChatCompletionChunk>, orchestrator::Error>>),
 }
 
 impl From<ChatCompletion> for ChatCompletionsResponse {
@@ -230,8 +230,8 @@ pub struct ChatCompletionsRequest {
     /// If set, partial message deltas will be sent, like in ChatGPT.
     /// Tokens will be sent as data-only server-sent events as they become available,
     /// with the stream terminated by a data: [DONE] message.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub stream: Option<bool>,
+    #[serde(default)]
+    pub stream: bool,
     /// Options for streaming response. Only set this when you set stream: true.
     #[serde(skip_serializing_if = "Option::is_none")]
     pub stream_options: Option<StreamOptions>,
@@ -544,10 +544,10 @@ pub struct ChatCompletion {
     /// This field is only included if the `service_tier` parameter is specified in the request.
     #[serde(skip_serializing_if = "Option::is_none")]
     pub service_tier: Option<String>,
-    /// Result of running different guardrail detectors
+    /// Detections
     #[serde(skip_serializing_if = "Option::is_none")]
     pub detections: Option<ChatDetections>,
-    /// Optional warnings
+    /// Warnings
     #[serde(default, skip_serializing_if = "Vec::is_empty")]
     pub warnings: Vec<OrchestratorWarning>,
 }
@@ -556,7 +556,7 @@ pub struct ChatCompletion {
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct ChatCompletionChoice {
     /// The index of the choice in the list of choices.
-    pub index: usize,
+    pub index: u32,
     /// A chat completion message generated by the model.
     pub message: ChatCompletionMessage,
     /// Log probability information for the choice.
@@ -612,7 +612,7 @@ pub struct ChatCompletionTopLogprob {
 }
 
 /// Represents a streamed chunk of a chat completion response returned by model, based on the provided input.
-#[derive(Debug, Clone, Serialize, Deserialize)]
+#[derive(Debug, Default, Clone, Serialize, Deserialize)]
 pub struct ChatCompletionChunk {
     /// A unique identifier for the chat completion. Each chunk has the same ID.
     pub id: String,
@@ -633,6 +633,12 @@ pub struct ChatCompletionChunk {
     pub service_tier: Option<String>,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub usage: Option<Usage>,
+    /// Detections
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub detections: Option<ChatDetections>,
+    /// Warnings
+    #[serde(default, skip_serializing_if = "Vec::is_empty")]
+    pub warnings: Vec<OrchestratorWarning>,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
@@ -711,7 +717,7 @@ pub struct OpenAiError {
 }
 
 /// Guardrails detection results.
-#[derive(Debug, Clone, Serialize, Deserialize)]
+#[derive(Debug, Default, Clone, Serialize, Deserialize)]
 pub struct ChatDetections {
     #[serde(default, skip_serializing_if = "Vec::is_empty")]
     pub input: Vec<InputDetectionResult>,
@@ -722,7 +728,7 @@ pub struct ChatDetections {
 /// Guardrails detection result for application on input.
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct InputDetectionResult {
-    pub message_index: usize,
+    pub message_index: u32,
     #[serde(default, skip_serializing_if = "Vec::is_empty")]
     pub results: Vec<ContentAnalysisResponse>,
 }
@@ -730,7 +736,7 @@ pub struct InputDetectionResult {
 /// Guardrails detection result for application output.
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct OutputDetectionResult {
-    pub choice_index: usize,
+    pub choice_index: u32,
     #[serde(default, skip_serializing_if = "Vec::is_empty")]
     pub results: Vec<ContentAnalysisResponse>,
 }
@@ -738,7 +744,7 @@ pub struct OutputDetectionResult {
 /// Represents the input and output of detection results following processing.
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct DetectionResult {
-    pub index: usize,
+    pub index: u32,
     #[serde(default, skip_serializing_if = "Vec::is_empty")]
     pub results: Vec<ContentAnalysisResponse>,
 }
