+++ src/orchestrator/common/utils.rs
@@ -1,0 +1,67 @@
+use std::{collections::HashMap, sync::Arc};
+
+use crate::{
+    models::DetectorParams,
+    orchestrator::{Context, Error},
+};
+
+/// Slices chars between start and end indices.
+pub fn slice_codepoints(text: &str, start: usize, end: usize) -> String {
+    let len = end - start;
+    text.chars().skip(start).take(len).collect()
+}
+
+/// Applies masks to input text, returning (offset, masked_text) pairs.
+pub fn apply_masks(text: String, masks: Option<&[(usize, usize)]>) -> Vec<(usize, String)> {
+    match masks {
+        None | Some([]) => vec![(0, text)],
+        Some(masks) => masks
+            .iter()
+            .map(|(start, end)| {
+                let masked_text = slice_codepoints(&text, *start, *end);
+                (*start, masked_text)
+            })
+            .collect(),
+    }
+}
+
+pub fn get_chunker_ids(
+    ctx: &Arc<Context>,
+    detectors: &HashMap<String, DetectorParams>,
+) -> Result<Vec<String>, Error> {
+    detectors
+        .keys()
+        .map(|detector_id| {
+            let chunker_id = ctx
+                .config
+                .get_chunker_id(detector_id)
+                .ok_or_else(|| Error::DetectorNotFound(detector_id.clone()))?;
+            Ok::<String, Error>(chunker_id)
+        })
+        .collect::<Result<Vec<_>, Error>>()
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn test_apply_masks() {
+        let text = "I want this sentence. I don't want this sentence. I want this sentence too.";
+        let masks: Vec<(usize, usize)> = vec![(0, 21), (50, 75)];
+        let text_with_offsets = apply_masks(text.into(), Some(&masks));
+        let expected_text_with_offsets = vec![
+            (0, "I want this sentence.".to_string()),
+            (50, "I want this sentence too.".to_string()),
+        ];
+        assert_eq!(text_with_offsets, expected_text_with_offsets)
+    }
+
+    #[test]
+    fn test_slice_codepoints() {
+        let s = "Hello world";
+        assert_eq!(slice_codepoints(s, 0, 5), "Hello");
+        let s = "哈囉世界";
+        assert_eq!(slice_codepoints(s, 3, 4), "界");
+    }
+}
